[ADD] springy text hover effect
